Directory bin/ removed
The directory used to contain code to create auxiliary files with event weights. The most of this code has been moved to PECLocalFwk
framework. The only remaining portion is PDF reweighting, which is not used in the analysis at the moment. When the analysis is ready
to incorporate it, the directory will probably be recovered from "initial" tag, and the code will be updated.