CMSSW tags have been updated to PAT recommendations
The tags have been updated according to [1]. Also LumiDB tag is updated to the latest one. Other tags (for b-tagging, MET corrections, etc.) have not been checked and and might also require an update w.r.t. to release versions. The code is compiled without errors, but it have not been run.
[1] https://twiki.cern.ch/twiki/bin/view/CMSPublic/SWGuidePATReleaseNotes52X#V08_09_62