Return jet pile-up ID
Introduce jet pile-up ID again. Previous time it was deemed unuseful and
removed in commit e5170ff4bca600d1490debdc0c59c28ab683c532. Now the
continious pile-up discirminators are not stored. Instead, only boolean
information on whether a jet passes each supported working point of the
three provided algorithms is saved. It allows to compress all
information in one short per jet, which is affordable even if the
pile-up ID will not be used in an analysis.

This is a draft implementation. Input tags to access value maps for
pile-up ID are hard-coded whereas they should be configurable. The
output is swamped by TMVA warnings about jets below 20 GeV/c as the
"full" pile-up discriminator is not trained for them. And I am still
waiting for the experts to confirm that the correction I had to do to
the default setup make sense.