Remove access to JEC and JER
Since recent versions of PECFwk are capable of applying JEC and JER on
the fly, there is no more need to store corresponding factors for jet
momenta, and they are dropped. In case of MC it frees up 4 floats per
jet.

Note that this commit makes the PlainEventContent plugin store only jets
that pass the cuts on raw or (defaultly) corrected pt; it does not store
all jets that might pass these cuts due to a variation of JEC or JER as
it used to be before. It must be considered when the pt cuts are
specified.