Remove trigger-emulating preselection for MVA electron ID
Hopefully, it will not be needed in Run II (and if it will, the
selection will be different anyway). Along with the trigger-emulating
preselection, remove correction for the problem in ECAL-based isolation.

Indices of the selection bits filled in pec::Electron have not been
changed, hence bit #1 is left unset.