Update jet part to miniAOD
There are no C++ or Python errors, but I haven't run the code. It's
likely that there will be problems when calculating the pull angle since
jet constituents are utilised for it.

I remove the block related to jet PU ID in ObjectsDefinitions_cff.py
completely. Mass of the secondary vertex is precalculated for miniAOD
and is thus access in a very different (much simpler) way then before.