Update electrons and muons for miniAOD
There are no C++ and Python errors, but I haven't run PlainEventContent.

Definition of leptons to be stored in tuples is likely to evolve in
future (will probably need to apply some loose ID and/or isolation
selection). For both electrons and muons delta-beta corrected isolation
with a scale factor of 0.5 is used; I suspect it's just a guess, and the
definition of the isolation will evolve.

I store electron ID decisions in bits for user-defined selections.
Dedicated fields for ID in pec::Electron are not used. Probably, I
should introduce a bit mask in the class for the ID specifically so that
I don't occupy too many bits meant for additional selections.