Align access to MET with miniAOD
Now all MET variations can be accessed from pat::MET, and there is no
need to produce them manually in the python configuration. Because of
this, the DefineMETs function is removed.

The MET used is type-I corrected. There are no type-0 and phi-modulation
corrections at the moment. Systematical variations of MET are not added
in this commit.