Put back loose event selection
On a ttbar sample efficiency of the selection in the electron channel is
larger than in the muon channel, which is suspecious. Probably, in the
electron channel there is a significant contribution of fake electrons
since no ID or isolation requirements are applied to them. In future it
might be then needed to apply some loose quality selection for the
electrons in order to limit the file sizes.

The selection on jects currently does not account for variations in JEC
or JER. This will have to be corrected in the future.