Accept 21 as gluon codes in PDF information
With Pythia 6 gluons in the initial state (as given in
GenEventInfoProduct::PDF) were encoded with zeros instead of their PDG
ID codes. In Pythia 8 samples the PDG ID code is used instead, which
caused PlainEventContent to crash (with an exception from
GeneratorInfo::SetPdfId). GeneratorInfo is updated to accept both codes.

GeneratorInfo::PdfId now returns the PDG ID codes for gluons.