Draft update of HardInteractionInfo for Pythia 8 with miniAOD
Change logic of the plugin: after this commit it does not simply save
all "status-3" particles (there is no such thing in Pythia 8), but
attempts to identify the initial and final states of the hard
interaction and also store selected interesting particles from the final
state (some candidates are H, t, W) and their decay products.

This plugin provides a draft implementation. Genetic information is not
stored at the moment. Comments on large chunks of code are missing and
will be added in one of later commits.

This addresses issue #41 and issue #43.