Disable minifloats for the time being
The philosophy behind the packages has changed, and now larger output
files can be tolerated. It is thus better to keep standard
floating-point implementation for the sake of ease of inspection of the
files in a TBrowser. Avoiding conversion should also result in a faster
reading.

At the moment minifloats are only disabled with the dedicated flag. Will
eliminate them completely later, when it is clear that they are not
needed (for instance, would like to check how much do generator-level
weights increase the file size). Documentation in PEC classes in what
concerns representation with minifloats is not correct after this commit
and must be updated if minifloats are really dropped.