Reapply JEC
This addresses issue #64.

Tested the effect of reapplying the corrections on MiniAODv2 simulation
but observed no differences in jet momenta. Exploited dataset was
re-MiniAOD'ed using the global tag 74X_mcRun2_asymptotic_v2. The
relevant JEC payload included there is Summer15_25nsV2_MC_AK4PFchs [1],
while JEC are reapplied using the V6 payload. According to wiki [2],
central values of MC corrections in these two payloads are identical
(although uncertainties are different). Thus, it is expected that
reapplication of JEC does not have effect on the current setup. To be
checked if JEC uncertainties are affected.

[1] https://cms-conddb.cern.ch/browser/#list/Prod/gts/74X_mcRun2_asymptotic_v2
[2] https://twiki.cern.ch/twiki/bin/viewauth/CMS/JECDataMC?rev=98#Recommended_for_MC