Enable storing of several MVA-based IDs for electrons
This will be useful if at some point need to save also non-triggering
ID. However, they are stored in an array, whose size must be known at
compile time. This is quite inconvenient, but didn't want to add a long
int per electron just to store an std::vector with IDs.