Apply some of filters recommended for analyses involving MET
Only add filters that can be rerun from MiniAOD or whose decisions are
stored as trigger results. Filters based on event ID will be introduced
later. The lists of events to be excluded can be very large, and for
this reason evaluation of MET-related filters is moved to after the
loose event selection.

This addresses issue #53.