Do not apply selection on jet ID but simply store its results as ID bit
With recent updates in jet ID and the HF sage it would be safer to be
able to check performance of jet ID rather than applying it blindly.
Since its efficiency is supposed to be more than 99% for real jets, and
noise does not contribute too much, the change should not increase the
file size any significantly.