Rename module patJetCorrFactorsReapplyJEC to avoid conflicts with MET
MET tool uses a similar module with the same name. Eventually might
consider changing the running schedule so that the same module is reused
for both jets and MET (and, maybe, jets with reapplied JEC are taken
from the MET tool). But for the time being separate the two modules for
the sake of safety.

This addresses issue #71.