Update style of Python code to comply with PEP8
Mostly changes code in the python/ directory while those in the
directory scripts/ is only superficially updated. Only the program code
is touched while comments and doc strings are not modified for the
moment.

There are few cases when recommendations of PEP8 are deliberately
violated. When a CMSSW plugin is configured, the name of its C++ class
is written on the first line and keyword arguments are indented by one
level (the standard recommends that either the first argument is
written on the second line with the same indentation as others or that
arguments in the subsequent lines are aligned to the first symbol of
the first argument written on the first line). This is done to emphasize
the importance of the C++ class name since just knowing if it's an
EDAnalyzer or EDProducer gives only a little information.

In the keyword arguments in configuration of CMSSW plugins (and only
there) the '=' signs are surrounded by spaces in contrast to what the
standard suggests. The motivation is that both the keyword names and
their values tend to be long (especially since the type needs to be
specified), and additional spaces improve readability ("Readability
counts" -- The Zen of Python).

CMSSW modules are allowed to be imported in the middle of code. They are
more like additional pieces of configuration rather than modules
extending functionality. Conceptually, they are similar to
process.load(...), which can be executed at any moment. Also keep in
mind that import might be conditional, depending on run-time
configuration (e.g. the value of the runOnData flag).

The maximal line width for code is set to 99 symbols. As recommended in
PEP 8, comments and docstrings are limited to 72 symbols though.
However, these limits do not apply to URLs and hard-coded file names.

This commit addresses issue #67.