Polish PEC classes from C++ side
There was a bit of a mess with defaulted methods. It is sufficient to
define a defaulted method in a base class, with no need to do it again
in every derived class. Because of this all defaulted copy constructors
and assignment operators in derived classes have been removed.

A similar consideration applies to destructors: if a base class has a
virtual destructor (which can be defaulted), compiler will generate
trivial virtual destructors for all inheriting classes. Moreover, a
default destructor is noexcept(true). As a result, there is no need to
declare destructors in inheriting classes explicitly.

When a user-defined destructor is provided (even defaulted one),
compiler does not generate implicit move functions. However, they are
useful in objects inheriting from pec::Candidate. To address this,
provide a defaulted move constructor for pec::Candidate. Inheriting
classes will have compiler-generated move constructors.

In addition, constructors with no parameters are turned to noexcept
since they are never expected to throw.