Update configuration for jets
Recompute PU jet ID on the fly, as it was done in 76X. In MiniAOD80Xv2
it will be available in userFloats, but in v1 it provides an outdated
ID. Will need to move to userFloats when update to MiniAOD80Xv2.

The set-up to override JER factors in global tag is removed, and thus
the package will use default factors included in the global tag. They
are for 74X [1] and will need to be updated eventually. Factors for 80X
are not available yet, of course.
[1] https://cms-conddb.cern.ch/cmsDbBrowser/list/Prod/gts/80X_mcRun2_asymptotic_2016_miniAODv2