Make LHEEventWeights more flexible about structure of LHE header
Files of private H/A->tt datasets contain lines like
  <weight MUF="1" MUR="1" PDF="263000" id="1"> mur=1 muf=1 </weight>
in the LHE header. In past the only parameter of the <weight> tag was
"id", while now other parameters are included. Not sure if this is a
change in the production code or a bug in the private production. All
the same, this commit allows arbitrary paramters in this tag.