Rearrange access to LHE event record
Commit [1] introduced a problem, which was hidden because
LHEEventProduct was read by default. After this commit these data must
be read to access process ID and not just when alternative LHE-level
weights need to be stored.
[1] 024e48a1e9ab0803191931fe4e9e33d92fa5d445

Now PECGenerator reads LHEEventProduct by default. In case of samples
not based on LHE files this must be disabled explicitly by giving an
empty label to be plugin.

This commit also renames configuration parameter saveLHEWeightVars ->
saveAltLHEWeights and improves relevant documentation.