Update scripts to process CRAB tasks
Remove script checkCrabRes.py as it is not needed with CRAB 3.

Heavily update script mergeCrabRes.py. Many parts have been rewritten
impoving style and clarity. It now verifies that the number of events in
the merged files is the same as in input files. The output files are not
placed in the current directory by default.