Update default input files and review jets and MET
Still use ICHEP JEC since newer corrections are not available in a
global tag yet. Definitions of jets and MET and MET filters seem to be
up to date, apart from jet ID, which will be updated in a different
commit.

Do not attempt to update electron and photon ID for the moment because
the corresponding recipes are expected to be updated soon.