In crabSubmit.py retry job submission if failed due to a server problem
Server problems are usually transient, and it is sufficient to retry
submission immediately. This makes the script to take up to three
attempts before giving up a task.

Colours of the output are also improved. For best performance terminal
should support 256 colours. When using with GNU screen, one might need
to add
  export TERM=xterm-256color
to .bash_profile so that this variable is picked up when screen starts.