Use saved flags to identify events with spurious muons
This commit actually fixes a bug. The collection of muons in the
re-MiniAOD samples have been cleaned from "bad" candidates, and as a
result the standalone Giovanni's filters do not work anymore. Filtering
based on the saved precomputed flags is now the only way to reject
events with duplicate or bad muons.

This fixes #96.