Do not filter on bad and spurious muons but store the flags instead
The bad muon flag rejects about 1% of events, which is not negligible.
Because of this should prefer to store the flags to have freedom to
reject or to keep the affected events when running over PEC files.
Hopefully, fixes in MET will allow to include these events in the
analysis.