Recompute MET corrections for ECAL gain switch
In data recompute MET correction due to ECAL gain switch because in rare
cases MET in slimmedMETsMuEGClean in MiniAOD the correction has some
issues.

In a test run see differences in MET with respect to one in MiniAOD only
in single events. Although ran over a file from SingleMuon PD.