Store information needed to evaluate independent JEC uncertainties
The main problem is that jets down to 15 GeV need to be stored in order
to propagate variation in JEC into MET. The cut in jet definition is
lowered, which increases the number of stored jets by about a factor of
two in data.

Also store MET with undone T1 correction from jets. This is done using
the actual correction that is applied to stored MET by the standard MET
tool. The resulting MET with undone correction coinsides with the raw
MET in both data and simulation. It is expected in simulation. In data
the real raw PF MET should be different because of the additional
corrections due to spurious muons and ECAL gain switch. But in reality
PATMETSlimmer is configured to store MET with this additional
corrections as the raw MET.