Clean up for 2017 data and update input files and global tags
Drop all references to 2016 legacy re-reconstruction. When eventually
MiniAOD for 2016 data is produced in 94X, it will be made from the
legacy re-reconstruction, so there is no need to support old prompt
data.

Old jet corrections are used at the moment (and not sure which ones will
be picked up for data) since they are the only ones available.