Do not recorrect jets and recompute missing pt
In define_jets function do not recorrect jets unless requested. This
used to be done as this was the only way to compute DeepCSV b tags, but
now they are stored in MiniAOD. Consequently, also drop their
computation even when jets are recorrected.

Recorrection of jets is switched off by default, and then missing pt can
also be read directly from MiniAOD, which is what is done now. As a
result, the object that defines the type 1 correction is not available
anymore. However, it was not really used, so it is dropped from the
output PEC file.