Only store LHE weights for scale variations
Allow selecting which LHE weights to store. In the Python configuration
only enable storing of the weights for the scale variations. In order to
save also selected PDF weights, the configuration will need to be
updated.

The selection applies also to mean values of the alternative weights
stored in plugin EventCounter. The name of the corresponding output
branch is changed to indicate that these are LHE weights.

This fixes #86.